requirements.txt
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.12
DateTime==4.4
docopt==0.6.2
idna==3.3
pipreqs==0.4.11
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
six @ file:///AppleInternal/BuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-110/six-1.15.0-py2.py3-none-any.whl
soupsieve==2.3.2.post1
urllib3==1.26.9
wget==3.2
yarg==0.1.9
zope.interface==5.4.0
